Rewrite B9 blog index to v2.2 spec
- Uniform 3-col grid (no featured hero post), newest-first sort
- Fixed category filters (RSE in Practice / Guidance and Policy / Our Work)
  with disabled state for empty categories, topic filter row
- New .blog-index-card component with 16:9 featured image, category
  eyebrow, excerpt, author/date meta, topic labels (capped at 3)
- Mobile horizontal-scroll chip track with fade gradient
- ARIA toolbar keyboard nav, roving tabindex, live region counts
- Load More pagination gated behind 24-post threshold
- URL state sync for filters (?category=…&topic=…)
- JSON-LD Blog schema with blogPost[] entries
- Page-scoped styles; remove 410 lines of stale v1.0 B9 CSS from
  tailor-site-v2.css that were overriding Astro scoped styles
- Fix formatDate() to short month format matching C4
- Add DOMContentLoaded fallback for Safari compatibility

+++ src/lib/blog-helpers.ts
@@ -6,7 +6,8 @@ import type { BlockObjectResponse } from '@notionhq/client/build/src/api-endpoin
 import type { BlogPost } from './types';
 
 /**
- * Format an ISO date string as British long form — e.g. "11 April 2026".
+ * Format an ISO date string as British short form — e.g. "9 Apr 2026".
+ * Matches the C4 blog post detail format exactly.
  * Returns null for null input so callers can `&&`-guard the output.
  */
 export function formatDate(iso: string | null): string | null {
@@ -15,7 +16,7 @@ export function formatDate(iso: string | null): string | null {
   if (Number.isNaN(date.getTime())) return null;
   return date.toLocaleDateString('en-GB', {
     day: 'numeric',
-    month: 'long',
+    month: 'short',
     year: 'numeric',
   });
 }
